refactor: Centralize Tool, ContentPart, and OAuth token types in provider -- go-llm-router (10)
