+++ internal/agents/oauth/copilot/oauth.go
@@ -10,6 +10,8 @@ import (
 
 	"github.com/pardnchiu/go-pkg/filesystem/keychain"
 	go_pkg_http "github.com/pardnchiu/go-pkg/http"
+
+	"github.com/pardnchiu/agenvoy/internal/agents/provider"
 )
 
 const (
@@ -22,13 +24,6 @@ const (
 var (
 	errAuthorizationPending = fmt.Errorf("authorization pending") // * pre declare error for ensuring padding wont cause login exit
 )
-
-type Token struct {
-	AccessToken string    `json:"access_token"`
-	TokenType   string    `json:"token_type"`
-	Scope       string    `json:"scope"`
-	ExpiresAt   time.Time `json:"expires_at"`
-}
 
 type DeviceCode struct {
 	DeviceCode      string `json:"device_code"`
@@ -45,7 +40,7 @@ type GopilotAccessToken struct {
 	Error       string `json:"error"`
 }
 
-func Load() (*Token, error) {
+func Load() (*provider.CopilotToken, error) {
 	raw := keychain.Get(tokenKey)
 	// ! agenvoy.copilot.token will deprecated in v1.*.*
 	if raw == "" {
@@ -54,7 +49,7 @@ func Load() (*Token, error) {
 	if raw == "" {
 		return nil, nil
 	}
-	var t Token
+	var t provider.CopilotToken
 	if err := json.Unmarshal([]byte(raw), &t); err != nil {
 		return nil, fmt.Errorf("json.Unmarshal: %w", err)
 	}
@@ -75,7 +70,7 @@ func ClearToken() error {
 	return err
 }
 
-func LoginWithCallback(ctx context.Context, onCode func(*DeviceCode)) (*Token, error) {
+func LoginWithCallback(ctx context.Context, onCode func(*DeviceCode)) (*provider.CopilotToken, error) {
 	code, _, err := go_pkg_http.POST[DeviceCode](ctx, nil, deviceCodeAPI,
 		map[string]string{},
 		map[string]any{
@@ -92,7 +87,7 @@ func LoginWithCallback(ctx context.Context, onCode func(*DeviceCode)) (*Token, e
 	interval := time.Duration(code.Interval) * time.Second
 	deadline := time.Now().Add(time.Duration(code.ExpiresIn) * time.Second)
 
-	var token *Token
+	var token *provider.CopilotToken
 	client := &http.Client{Timeout: 30 * time.Second}
 	for time.Now().Before(deadline) {
 		select {
@@ -114,7 +109,7 @@ func LoginWithCallback(ctx context.Context, onCode func(*DeviceCode)) (*Token, e
 	return nil, fmt.Errorf("device code expired")
 }
 
-func getAccessToken(ctx context.Context, client *http.Client, deviceCode string) (*Token, error) {
+func getAccessToken(ctx context.Context, client *http.Client, deviceCode string) (*provider.CopilotToken, error) {
 	accessToken, _, err := go_pkg_http.POST[GopilotAccessToken](ctx, client, oauthAccessTokenAPI,
 		map[string]string{},
 		map[string]any{
@@ -128,7 +123,7 @@ func getAccessToken(ctx context.Context, client *http.Client, deviceCode string)
 
 	switch accessToken.Error {
 	case "":
-		token := &Token{
+		token := &provider.CopilotToken{
 			AccessToken: accessToken.AccessToken,
 			TokenType:   accessToken.TokenType,
 			Scope:       accessToken.Scope,
