+++ internal/agents/oauth/grok/login.go
@@ -12,9 +12,11 @@ import (
 	"net/url"
 	"strings"
 	"time"
+
+	"github.com/pardnchiu/agenvoy/internal/agents/provider"
 )
 
-func LoginWithCallback(ctx context.Context, onURL func(string)) (*StoredToken, error) {
+func LoginWithCallback(ctx context.Context, onURL func(string)) (*provider.GrokToken, error) {
 	b := make([]byte, 32)
 	if _, err := rand.Read(b); err != nil {
 		return nil, fmt.Errorf("failed to generate PKCE verifier: %w", err)
@@ -59,7 +61,7 @@ func LoginWithCallback(ctx context.Context, onURL func(string)) (*StoredToken, e
 	}
 }
 
-func exchangeCode(ctx context.Context, code, verifier, challenge string) (*StoredToken, error) {
+func exchangeCode(ctx context.Context, code, verifier, challenge string) (*provider.GrokToken, error) {
 	form := url.Values{
 		"grant_type":            {"authorization_code"},
 		"code":                  {code},
@@ -96,7 +98,7 @@ func exchangeCode(ctx context.Context, code, verifier, challenge string) (*Store
 		expiry = time.Now().Add(3600 * time.Second)
 	}
 
-	token := &StoredToken{
+	token := &provider.GrokToken{
 		AccessToken:  raw.AccessToken,
 		RefreshToken: raw.RefreshToken,
 		ExpiresAt:    expiry,
