+++ internal/agents/oauth/grok/type.go
@@ -1,7 +1,5 @@
 package oauthGrok
 
-import "time"
-
 const (
 	tokenKey    = "GROK_OAUTH_TOKEN"
 	clientID    = "b1a00492-073a-47ea-816f-4c329264a828"
@@ -10,16 +8,6 @@ const (
 	redirectURI = "http://127.0.0.1:56121/callback"
 	scopes      = "openid profile email offline_access grok-cli:access api:access"
 )
-
-type StoredToken struct {
-	AccessToken  string    `json:"access_token"`
-	RefreshToken string    `json:"refresh_token"`
-	ExpiresAt    time.Time `json:"expires_at"`
-}
-
-func (t *StoredToken) expired() bool {
-	return time.Now().After(t.ExpiresAt.Add(-60 * time.Second))
-}
 
 type oauthTokenResponse struct {
 	AccessToken  string `json:"access_token"`
