+++ internal/agents/provider/compat/send.go
@@ -5,11 +5,10 @@ import (
 	"fmt"
 
 	"github.com/pardnchiu/agenvoy/internal/agents/provider"
-	toolTypes "github.com/pardnchiu/agenvoy/internal/tools/types"
 	go_pkg_http "github.com/pardnchiu/go-pkg/http"
 )
 
-func (a *Agent) Send(ctx context.Context, messages []provider.Message, tools []toolTypes.Tool) (*provider.Output, error) {
+func (a *Agent) Send(ctx context.Context, messages []provider.Message, tools []provider.Tool) (*provider.Output, error) {
 	chatAPI := a.baseURL + "/chat/completions"
 
 	headers := map[string]string{
