--- internal/agents/provider/config.go
@@ -1,12 +1,0 @@
-package provider
-
-type Config struct {
-	Model   string
-	APIKey  string
-	Token   any
-	BaseURL string
-
-	// * cloudflare
-	AccountID string
-	GatewayID string
-}
