--- internal/agents/provider/output.go
@@ -1,67 +1,0 @@
-package provider
-
-import "encoding/json"
-
-type Message struct {
-	Role             string     `json:"role"`
-	Content          any        `json:"content,omitempty"`
-	ReasoningContent string     `json:"reasoning_content,omitempty"`
-	ToolCalls        []ToolCall `json:"tool_calls,omitempty"`
-	ToolCallID       string     `json:"tool_call_id,omitempty"`
-}
-
-type ToolCall struct {
-	ID       string `json:"id"`
-	Type     string `json:"type"`
-	Function struct {
-		Name      string `json:"name"`
-		Arguments string `json:"arguments"`
-	} `json:"function"`
-	ThoughtSignature string `json:"thought_signature,omitempty"` // Gemini thinking models
-}
-
-type Output struct {
-	Choices []OutputChoices `json:"choices"`
-	Usage   Usage           `json:"usage"`
-	Error   *struct {
-		Message string      `json:"message"`
-		Type    string      `json:"type"`
-		Code    json.Number `json:"code"`
-	} `json:"error,omitempty"`
-}
-
-type OutputChoices struct {
-	Message      Message `json:"message"`
-	Delta        Message `json:"delta"`
-	FinishReason string  `json:"finish_reason,omitempty"`
-}
-
-type Usage struct {
-	Input       int `json:"input_tokens"`
-	Output      int `json:"output_tokens"`
-	CacheCreate int `json:"cache_creation_input_tokens,omitempty"`
-	CacheRead   int `json:"cache_read_input_tokens,omitempty"`
-}
-
-func (u *Usage) UnmarshalJSON(data []byte) error {
-	var raw struct {
-		InputTokens              int `json:"input_tokens"`
-		OutputTokens             int `json:"output_tokens"`
-		PromptTokens             int `json:"prompt_tokens"`
-		CompletionTokens         int `json:"completion_tokens"`
-		CacheCreationInputTokens int `json:"cache_creation_input_tokens"`
-		CacheReadInputTokens     int `json:"cache_read_input_tokens"`
-		PromptTokensDetails      struct {
-			CachedTokens int `json:"cached_tokens"`
-		} `json:"prompt_tokens_details"`
-	}
-	if err := json.Unmarshal(data, &raw); err != nil {
-		return err
-	}
-	cached := raw.CacheReadInputTokens + raw.PromptTokensDetails.CachedTokens
-	u.Input = raw.InputTokens + raw.PromptTokens - raw.PromptTokensDetails.CachedTokens
-	u.Output = raw.OutputTokens + raw.CompletionTokens
-	u.CacheCreate = raw.CacheCreationInputTokens
-	u.CacheRead = cached
-	return nil
-}
