+++ internal/session/config/bot/bot.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/pardnchiu/agenvoy/configs"
 	"github.com/pardnchiu/agenvoy/internal/filesystem"
+	"github.com/pardnchiu/agenvoy/internal/session/config"
 )
 
 const (
@@ -89,7 +90,12 @@ func GetModel(sessionID string) (model, reasoning string) {
 		model = DefaultModel
 	}
 	if reasoning == "" {
-		reasoning = DefaultReasoning
+		cfg, err := config.Load()
+		if err == nil && cfg.ReasoningLevel != "" {
+			reasoning = cfg.ReasoningLevel
+		} else {
+			reasoning = DefaultReasoning
+		}
 	}
 	return model, reasoning
 }
